hash_id: Rather than adding the IDs, combine into one uint64_t.
- hash_id should take a uint64_t argument, rather than unsigned.

- Instead of adding them or hashing them separately and combining,
  pack both into the uint64_t argument for hash_id, since each is
  a 32-bit ID. Further experimentation supports that this has
  better collision behavior.

+++ include/adt/hash.h
@@ -17,7 +17,7 @@
 
 SUPPRESS_EXPECTED_UNSIGNED_INTEGER_OVERFLOW()
 static __inline__ uint64_t
-hash_id(unsigned id)
+hash_id(uint64_t id)
 {
 	/* xorshift* A1(12,25,27),
 	 * from http://vigna.di.unimi.it/ftp/papers/xorshift.pdf */
